add multi part send and receive (#2)
* add multi part send and receive

This reduces coding effort for working with multipart messages. And it's
also more efficient, especially for messages with a large number of parts.

* handle EINTR during send/receive-multipart

It's possible that during a multipart send/receive the zmq operation is
interrupted by a signal. In this case the multipart should continue
trying, so that a multi part message is either fully received, or not at
all (if the recv/send is the first operation).

* multipart send/recv: replace errno by zmq_errno() to support NON POSIX systems

* multipart send/recv: added timeout test